requirements.txt
nltk >= 3.2.1 
numpy >= 1.6.1
pandas >= 0.18.0
scipy >= 0.9.0
scikit-learn >= 0.17.1
python-dotenv == 0.6.0
coverage == 4.2
